--- src/pages/home/components/HomeLayout/PageContents/index.ts
@@ -1,1 +1,0 @@
-export { default as PageContents } from './PageContents';
