+++ src/pages/home/components/HomeLayout/PageRouter/index.ts
@@ -1,0 +1,1 @@
+export { default as PageRouter } from './PageRouter';
